+++ retrofit/lib/dio.dart
@@ -1,5 +1,5 @@
-import 'package:meta/meta.dart';
 import 'package:dio/dio.dart';
+import 'package:meta/meta.dart';
 
 /// Extra data that will be passed to dio's request, response, transformer and interceptors.
 @immutable
@@ -28,12 +28,14 @@ class SendProgress {
 @immutable
 class DioResponseType {
   final ResponseType responseType;
+
   const DioResponseType(this.responseType);
 }
 
 class HttpResponse<T> {
   final T data;
   final Response response;
+
   HttpResponse(this.data, this.response);
 }