+++ retrofit/lib/retrofit.dart
@@ -1,2 +1,2 @@
-export 'http.dart';
 export 'dio.dart';
+export 'http.dart';